Fix: Layout and rendering issues in orcamentos calculator
- Allow decimal input (step=0.1) for dimensions (largura, altura, sangria)
- Convert backend parsing from int() to float() to support decimal values
- Fix template variable rendering by removing newlines from Django tags
- Replace SVG pattern with individual rectangle rendering for accurate visualization
- Center grid layout on paper using calculated offsets
- Add hx-swap='outerHTML' to prevent layout overlap issues
- Move approval visualization to fixed position between form sections
- Fix margin indicator stroke width and visual alignment

+++ orcamentos/views.py
@@ -21,9 +21,9 @@ class NovoOrcamentoView(LoginRequiredMixin, TemplateView):
 def htmx_calcular_aproveitamento(request):
     try:
         # 1. Coleta dados do Request
-        largura_final = int(request.GET.get('item-largura_final_mm') or 0)
-        altura_final = int(request.GET.get('item-altura_final_mm') or 0)
-        sangria = int(request.GET.get('item-sangria_mm') or 0)
+        largura_final = float(request.GET.get('item-largura_final_mm') or 0)
+        altura_final = float(request.GET.get('item-altura_final_mm') or 0)
+        sangria = float(request.GET.get('item-sangria_mm') or 0)
         papel_id = request.GET.get('parte-papel')
         
         # 2. Validações básicas
